+++ tests/Unit/Builders/AngularDistance/FromStringTest.php
@@ -1,0 +1,75 @@
+<?php
+namespace MarcoConsiglio\Goniometry\Tests\Unit\Builders\AngularDistance;
+
+use MarcoConsiglio\Goniometry\Angle;
+use MarcoConsiglio\Goniometry\AngularDistance;
+use MarcoConsiglio\Goniometry\Builders\Angle\FromSexadecimal as AngleFromSexadecimal;
+use MarcoConsiglio\Goniometry\Builders\AngularDistance\FromSexadecimal;
+use MarcoConsiglio\Goniometry\Builders\AngularDistance\FromString;
+use MarcoConsiglio\Goniometry\Degrees;
+use MarcoConsiglio\Goniometry\Minutes;
+use MarcoConsiglio\Goniometry\Random\Generator\Angle as AngleGenerator;
+use MarcoConsiglio\Goniometry\Random\Generator\NegativeAngle as NegativeAngleGenerator;
+use MarcoConsiglio\Goniometry\Random\Generator\NegativeSexadecimal as NegativeSexadecimalGenerator;
+use MarcoConsiglio\Goniometry\Random\Generator\PositiveAngle as PositiveAngleGenerator;
+use MarcoConsiglio\Goniometry\Random\Generator\PositiveSexadecimal as PositiveSexadecimalGenerator;
+use MarcoConsiglio\Goniometry\Random\Generator\RelativeAngle as RelativeAngleGenerator;
+use MarcoConsiglio\Goniometry\Random\Validator\FloatValidator;
+use MarcoConsiglio\Goniometry\Random\Validator\NegativeSexadecimal as NegativeSexadecimalValidator;
+use MarcoConsiglio\Goniometry\Random\Validator\PositiveSexadecimal as PositiveSexadecimalValidator;
+use MarcoConsiglio\Goniometry\Random\Validator\RelativeSexadecimal as RelativeSexadecimalValidator;
+use MarcoConsiglio\Goniometry\Seconds;
+use MarcoConsiglio\Goniometry\SexadecimalAngularDistance;
+use MarcoConsiglio\Goniometry\SexadecimalDegrees;
+use MarcoConsiglio\Goniometry\SexagesimalDegrees;
+use MarcoConsiglio\Goniometry\Tests\TestCase;
+use MarcoConsiglio\Goniometry\Traits\WithAngleFaker;
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\UsesClass;
+use PHPUnit\Framework\Attributes\UsesTrait;
+
+#[CoversClass(FromString::class)]
+#[UsesClass(Angle::class)]
+#[UsesClass(AngularDistance::class)]
+#[UsesClass(AngleFromSexadecimal::class)]
+#[UsesClass(FromSexadecimal::class)]
+#[UsesClass(Degrees::class)]
+#[UsesClass(Minutes::class)]
+#[UsesClass(AngleGenerator::class)]
+#[UsesClass(NegativeAngleGenerator::class)]
+#[UsesClass(NegativeSexadecimalGenerator::class)]
+#[UsesClass(RelativeAngleGenerator::class)]
+#[UsesClass(FloatValidator::class)]
+#[UsesClass(NegativeSexadecimalValidator::class)]
+#[UsesClass(RelativeSexadecimalValidator::class)]
+#[UsesClass(Seconds::class)]
+#[UsesClass(SexadecimalAngularDistance::class)]
+#[UsesClass(SexadecimalDegrees::class)]
+#[UsesClass(SexagesimalDegrees::class)]
+#[UsesClass(PositiveAngleGenerator::class)]
+#[UsesClass(PositiveSexadecimalGenerator::class)]
+#[UsesClass(PositiveSexadecimalValidator::class)]
+#[UsesTrait(WithAngleFaker::class)]
+class FromStringTest extends TestCase
+{
+    public function test_create_from_string(): void
+    {
+        // Arrange
+        $angle = $this->randomAngle(precision: 3);
+        $sexadecimal = new SexadecimalAngularDistance($angle->toSexadecimalDegrees()->value);
+        $angular_distance = AngularDistance::createFromDecimal($sexadecimal);
+        $builder = new FromString((string) $angle);
+
+        // Act
+        $result = $builder->fetchData();
+        $actual = $result[0];
+
+        // Assert
+        $this->assertInstanceOf(SexagesimalDegrees::class, $result[0]);
+        $this->assertInstanceOf(SexadecimalAngularDistance::class, $result[1]);
+        $this->assertDegrees($angular_distance->degrees, $actual->degrees);
+        $this->assertMinutes($angular_distance->minutes, $actual->minutes);
+        $this->assertSeconds($angular_distance->seconds, $actual->seconds, 3);
+        $this->assertDirection($angular_distance->direction, $actual->direction);
+    }
+}